Add BuildCache
Split out file changes that require a re-porcesses vs a re-export.
Add BuildCache to cache file in memory.
Add the Listen gem to blow the BuildCache when a file changes

+++ lib/condenser/cache/file_store.rb
@@ -105,9 +105,7 @@ class Condenser::Cache
     private
 
     def safe_open(path, &block)
-      if File.exist?(path)
-        File.open(path, 'rb', &block)
-      end
+      File.open(path, 'rb', &block) if File.exist?(path)
     rescue Errno::ENOENT
     end
